src/backend/requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
requests==2.32.3
beautifulsoup4==4.12.3
pypdf==4.3.1
docx2txt==0.8
sqlalchemy==2.0.34
aiosqlite==0.20.0
llama-index-core==0.11.23
llama-index-vector-stores-faiss<0.3.0
llama-index-embeddings-openai==0.2.5
llama-index-llms-openai==0.2.16
faiss-cpu==1.9.0
langgraph==0.2.24
openai==1.47.0
pytest==8.3.3
httpx==0.27.2
urllib3==1.26.18
socksio==1.0.0
python-docx>=1.0.0
openpyxl>=3.1.0
python-pptx>=1.0.0
xlrd>=2.0.0
markdown>=3.4.0
html2text>=2024.0.0
chardet>=5.0.0
rank-bm25>=0.2.2
numpy>=1.24.0
fastapi-mcp>=0.4.0
